osd: split to osd.c and osd-classic.c
