+++ src/osd/meson.build
@@ -1,4 +1,5 @@
 labwc_sources += files(
   'osd.c',
+  'osd-classic.c',
   'osd-field.c',
 )
